reformatted file structure
