--- src/main/java/btestats/btestats/API/Router.java
@@ -1,11 +1,0 @@
-package btestats.btestats.API;
-
-public class Router {
-    /*
-    Create server to route rest api.
-    Port should be set in config.
-    REST API should be an optional arg in config. If disabled, do not start the rest api server.
-    1. function to start api
-    2. function to stop api
-     */
-}
