--- src/main/java/btestats/btestats/API/Routes/BlocksPlacedLeaderboard.java
@@ -1,7 +1,0 @@
-package btestats.btestats.API.Routes;
-
-public class BlocksPlacedLeaderboard {
-    /* TODO Implement
-    1. GET /blocksPlacedLeaderboard
-     */
-}
