--- src/main/java/btestats/btestats/API/Routes/LoginStreakLeaderboard.java
@@ -1,7 +1,0 @@
-package btestats.btestats.API.Routes;
-
-public class LoginStreakLeaderboard {
-        /* TODO Implement
-    1. GET /loginStreakLeaderboard
-     */
-}
