--- src/main/java/btestats/btestats/API/Routes/PercentCompletion.java
@@ -1,7 +1,0 @@
-package btestats.btestats.API.Routes;
-
-public class PercentCompletion {
-    /* TODO Implement
-    1. GET /percentCompletion
-     */
-}
