--- src/main/java/btestats/btestats/API/Routes/PlayerStats.java
@@ -1,7 +1,0 @@
-package btestats.btestats.API.Routes;
-
-public class PlayerStats {
-    /*
-    1. GET /playerStats
-     */
-}
