--- src/main/java/btestats/btestats/API/Routes/Region.java
@@ -1,7 +1,0 @@
-package btestats.btestats.API.Routes;
-
-public class Region {
-    /*
-    1. GET /region
-     */
-}
