--- src/main/java/btestats/btestats/Commands/GetBlocksPlacedLeaderboard.java
@@ -1,8 +1,0 @@
-package btestats.btestats.Commands;
-
-public class GetBlocksPlacedLeaderboard {
-    /* TODO Implement
-    Generates leaderboard of most blocks placed.
-
-     */
-}
