--- src/main/java/btestats/btestats/Commands/GetLoginStreakLeaderboard.java
@@ -1,7 +1,0 @@
-package btestats.btestats.Commands;
-
-public class GetLoginStreakLeaderboard {
-    /* TODO Implement
-    Generates login streak leaderboard
-     */
-}
