--- src/main/java/btestats/btestats/Commands/GetPercentCompletion.java
@@ -1,7 +1,0 @@
-package btestats.btestats.Commands;
-
-public class GetPercentCompletion {
-    /* TODO Implement
-    Generate percentage of map completed
-     */
-}
