--- src/main/java/btestats/btestats/Commands/GetRegion.java
@@ -1,7 +1,0 @@
-package btestats.btestats.Commands;
-
-public class GetRegion {
-    /* TODO Implement:
-    1. Gets region. If no region provided as an argument, gets region user is currently in.
-     */
-}
