--- src/main/java/btestats/btestats/Database/Bufferable.java
@@ -1,7 +1,0 @@
-package btestats.btestats.Database;
-
-import org.bukkit.Location;
-
-public interface Bufferable {
-    public abstract void flush();
-}
