--- src/main/java/btestats/btestats/Database/Validation.java
@@ -1,8 +1,0 @@
-package btestats.btestats.Database;
-
-public class Validation {
-    /*
-    Verify DB and command input is correct before continuing.
-    Ex: make sure uuid is valid through a function verifyUUID(uuid: string) -> string, where the function returns a trimmed and formatted uuid.
-     */
-}
