--- src/main/java/btestats/btestats/Events/AddBlockPlace.java
@@ -1,36 +1,0 @@
-package btestats.btestats.Events;
-import btestats.btestats.Database.Players;
-import btestats.btestats.Util.Blocks.BlockOwnerHistory;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockPlaceEvent;
-
-public class AddBlockPlace implements Listener {
-    private final Players playerDB;
-    private final BlockOwnerHistory blockOwnerHistory;
-
-    public AddBlockPlace(Players playerDB, BlockOwnerHistory blockOwnerHistory){
-        this.playerDB = playerDB;
-        this.blockOwnerHistory = blockOwnerHistory;
-    }
-
-    @EventHandler
-    private void onBlockPlace(BlockPlaceEvent e){
-        Block block = e.getBlock();
-        Player player = e.getPlayer();
-        String uuid = player.getUniqueId().toString();
-
-        String owner = blockOwnerHistory.get(block);
-
-        // If Block was placed by somebody else
-        if (owner != null && owner.equals(uuid)){
-            return;
-        }
-
-        blockOwnerHistory.set(block, uuid);
-
-        this.playerDB.updateBlocksPlaced(uuid, 1);
-    }
-}
