--- src/main/java/btestats/btestats/Events/BlockOwnerCollectionFlush.java
@@ -1,31 +1,0 @@
-package btestats.btestats.Events;
-
-import btestats.btestats.Database.BlockOwnerCollection;
-import btestats.btestats.Database.Parsers.PlayerData;
-import btestats.btestats.Database.Players;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.world.WorldSaveEvent;
-
-
-public class BlockOwnerCollectionFlush implements Listener {
-    /*
-     * Flushes block history into DB every world save
-     */
-
-    private final BlockOwnerCollection blockOwnerCollection;
-    private final Players player;
-    public BlockOwnerCollectionFlush(BlockOwnerCollection blockOwnerCollection, Players player) {
-        this.blockOwnerCollection = blockOwnerCollection;
-        this.player = player;
-    }
-
-
-
-    @EventHandler
-    public void onWorldSave( WorldSaveEvent event ) {
-        blockOwnerCollection.flush();
-        player.flush();
-    }
-
-}
