--- src/main/java/btestats/btestats/Events/UpdateLoginStreak.java
@@ -1,7 +1,0 @@
-package btestats.btestats.Events;
-
-public class UpdateLoginStreak {
-    /*TODO Implement:
-    1. For every login, check last login time. If it's not today, update login streak accordingly.
-     */
-}
