--- src/main/java/btestats/btestats/Seed/GenerateRegions.java
@@ -1,9 +1,0 @@
-package btestats.btestats.Seed;
-
-public class GenerateRegions {
-    public static void main(String[] args){
-        /* TODO Implement
-            Generate regions from OSM and insert into DB
-         */
-    }
-}
