+++ src/main/java/net/btenj/btestats/commands/GetBlocksPlacedLeaderboard.java
@@ -1,0 +1,8 @@
+package net.btenj.btestats.commands;
+
+public class GetBlocksPlacedLeaderboard {
+    /* TODO Implement
+    Generates leaderboard of most blocks placed.
+
+     */
+}
