+++ src/main/java/net/btenj/btestats/commands/GetLoginStreakLeaderboard.java
@@ -1,0 +1,7 @@
+package net.btenj.btestats.commands;
+
+public class GetLoginStreakLeaderboard {
+    /* TODO Implement
+    Generates login streak leaderboard
+     */
+}
