+++ src/main/java/net/btenj/btestats/commands/GetPercentCompletion.java
@@ -1,0 +1,7 @@
+package net.btenj.btestats.commands;
+
+public class GetPercentCompletion {
+    /* TODO Implement
+    Generate percentage of map completed
+     */
+}
