+++ src/main/java/net/btenj/btestats/commands/GetRegion.java
@@ -1,0 +1,7 @@
+package net.btenj.btestats.commands;
+
+public class GetRegion {
+    /* TODO Implement:
+    1. Gets region. If no region provided as an argument, gets region user is currently in.
+     */
+}
