+++ src/main/java/net/btenj/btestats/events/UpdateLoginStreak.java
@@ -1,0 +1,7 @@
+package net.btenj.btestats.events;
+
+public class UpdateLoginStreak {
+    /*TODO Implement:
+    1. For every login, check last login time. If it's not today, update login streak accordingly.
+     */
+}
