+++ src/main/java/net/btenj/btestats/storage/Bufferable.java
@@ -1,0 +1,5 @@
+package net.btenj.btestats.storage;
+
+public interface Bufferable {
+    public abstract void flush();
+}
