+++ src/main/java/net/btenj/btestats/storage/Regions.java
@@ -1,0 +1,16 @@
+package net.btenj.btestats.storage;
+
+public class Regions {
+    /*TODO Implement:
+    1. DB Collection name is "Regions" and an object will look like:
+    {
+        _id: mongo id,
+        "name": string,
+        bounds: region data (this should be automatically generated with terra),
+        complete: boolean
+    }
+    2. Create region, edit region, delete region function. When creating region, make sure regions don't overlap
+    3. % of map complete function -> returns region completed / total region.
+    Note: EVERY DB CALL MUST FIRST GO THROUGH VALIDATION
+     */
+}
