+++ src/main/java/net/btenj/btestats/storage/Validation.java
@@ -1,0 +1,8 @@
+package net.btenj.btestats.storage;
+
+public class Validation {
+    /*
+    Verify DB and command input is correct before continuing.
+    Ex: make sure uuid is valid through a function verifyUUID(uuid: string) -> string, where the function returns a trimmed and formatted uuid.
+     */
+}
